src/graphrag/others/GraphRAG-JayLZhou/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
anytree==2.12.1
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
autograd==1.7.0
backoff==2.2.1
bcrypt==4.3.0
beartype==0.18.5
beautifulsoup4==4.13.3
bitarray==3.1.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colbert==0.40
colbert-ai==0.2.21
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.3.2
deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
faiss-gpu==1.9.0.post1
fastapi==0.115.11
filelock==3.17.0
filetype==1.2.0
flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gensim==4.3.3
git-python==1.0.3
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.1
graspologic==3.4.1
graspologic-native==1.2.3
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
humanfriendly==10.0
hyppo==0.4.0
icalendar==6.1.1
idna==3.10
igraph==0.11.8
importlib-metadata==8.5.0
importlib-resources==6.5.2
itsdangerous==2.2.0
jina2==3.1.6
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
kubernetes==32.0.1
lazy-object-proxy==1.10.0
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index==0.12.23
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.23.post2
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-ollama==0.6.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-legacy==0.9.48.post4
llama-index-llms-ollama==0.5.3
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-chroma==0.4.1
llama-index-vector-stores-faiss==0.3.0
llama-parse==0.6.4.post1
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mauve-text==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.93
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.21.0
openai==1.65.4
orjson==3.10.15
overrides==7.7.0
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
posthog==3.19.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybind11==2.13.6
pydantic==2.10.6
pydantic-core==2.27.2
pyfiglet==1.0.2
pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pypdf==5.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rouge-score==0.1.2
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==3.4.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.38
starlette==0.46.1
statsmodels==0.14.4
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0